requirements.txt
# Core
streamlit
pandas
numpy
sqlalchemy
mysqlclient

# Machine Learning
scikit-learn
joblib
statsmodels

# Visualization & Maps
matplotlib
seaborn
plotly
folium
pydeck

# Utilities
python-dotenv
loguru
pytest
